Place DisplayContext directly in Cx
